requirements.txt
django==4.1.5
django-taggit==3.1.0
pillow==9.4.0
django-tinymce==3.5.0